Rename: 클래스명 변경 및 파일 이동에 따른 import 수정
Related to: #47
