+++ src/main/java/pulleydoreurae/careerquestbackend/community/service/CommentService.java
@@ -6,9 +6,9 @@ import org.springframework.stereotype.Service;
 
 import pulleydoreurae.careerquestbackend.auth.domain.entity.UserAccount;
 import pulleydoreurae.careerquestbackend.community.domain.entity.Comment;
-import pulleydoreurae.careerquestbackend.common.community.domain.dto.request.CommentRequest;
+import pulleydoreurae.careerquestbackend.community.domain.dto.request.CommentRequest;
 import pulleydoreurae.careerquestbackend.common.community.domain.entity.Post;
-import pulleydoreurae.careerquestbackend.common.community.repository.CommentRepository;
+import pulleydoreurae.careerquestbackend.community.repository.CommentRepository;
 import pulleydoreurae.careerquestbackend.common.community.service.CommentService;
 import pulleydoreurae.careerquestbackend.common.community.service.CommonCommunityService;
