+++ src/main/java/pulleydoreurae/careerquestbackend/community/service/PostService.java
@@ -6,10 +6,10 @@ import org.springframework.stereotype.Service;
 import pulleydoreurae.careerquestbackend.community.domain.entity.PostImage;
 import pulleydoreurae.careerquestbackend.community.domain.entity.PostViewCheck;
 import pulleydoreurae.careerquestbackend.common.community.domain.entity.Post;
-import pulleydoreurae.careerquestbackend.common.community.repository.PostImageRepository;
-import pulleydoreurae.careerquestbackend.common.community.repository.PostLikeRepository;
-import pulleydoreurae.careerquestbackend.common.community.repository.PostRepository;
-import pulleydoreurae.careerquestbackend.common.community.repository.PostViewCheckRepository;
+import pulleydoreurae.careerquestbackend.community.repository.PostImageRepository;
+import pulleydoreurae.careerquestbackend.community.repository.PostLikeRepository;
+import pulleydoreurae.careerquestbackend.community.repository.PostRepository;
+import pulleydoreurae.careerquestbackend.community.repository.PostViewCheckRepository;
 import pulleydoreurae.careerquestbackend.common.community.service.CommonCommunityService;
 import pulleydoreurae.careerquestbackend.common.community.service.PostService;
 import pulleydoreurae.careerquestbackend.common.service.FileManagementService;
