+++ src/test/java/pulleydoreurae/careerquestbackend/community/service/BasicCommonCommunityServiceTest.java
@@ -20,22 +20,22 @@ import org.springframework.security.core.userdetails.UsernameNotFoundException;
 
 import pulleydoreurae.careerquestbackend.auth.domain.entity.UserAccount;
 import pulleydoreurae.careerquestbackend.auth.repository.UserAccountRepository;
-import pulleydoreurae.careerquestbackend.common.community.domain.PostCategory;
-import pulleydoreurae.careerquestbackend.common.community.domain.dto.request.PostRequest;
-import pulleydoreurae.careerquestbackend.common.community.domain.dto.response.PostResponse;
+import pulleydoreurae.careerquestbackend.community.domain.PostCategory;
+import pulleydoreurae.careerquestbackend.community.domain.dto.request.PostRequest;
+import pulleydoreurae.careerquestbackend.community.domain.dto.response.PostResponse;
 import pulleydoreurae.careerquestbackend.community.domain.entity.Comment;
 import pulleydoreurae.careerquestbackend.community.domain.entity.Post;
 import pulleydoreurae.careerquestbackend.community.domain.entity.PostImage;
 import pulleydoreurae.careerquestbackend.community.domain.entity.PostLike;
 import pulleydoreurae.careerquestbackend.community.domain.entity.PostViewCheck;
-import pulleydoreurae.careerquestbackend.common.community.exception.CommentNotFoundException;
-import pulleydoreurae.careerquestbackend.common.community.exception.PostLikeNotFoundException;
-import pulleydoreurae.careerquestbackend.common.community.exception.PostNotFoundException;
-import pulleydoreurae.careerquestbackend.common.community.repository.CommentRepository;
-import pulleydoreurae.careerquestbackend.common.community.repository.PostImageRepository;
-import pulleydoreurae.careerquestbackend.common.community.repository.PostLikeRepository;
-import pulleydoreurae.careerquestbackend.common.community.repository.PostRepository;
-import pulleydoreurae.careerquestbackend.common.community.repository.PostViewCheckRepository;
+import pulleydoreurae.careerquestbackend.community.exception.CommentNotFoundException;
+import pulleydoreurae.careerquestbackend.community.exception.PostLikeNotFoundException;
+import pulleydoreurae.careerquestbackend.community.exception.PostNotFoundException;
+import pulleydoreurae.careerquestbackend.community.repository.CommentRepository;
+import pulleydoreurae.careerquestbackend.community.repository.PostImageRepository;
+import pulleydoreurae.careerquestbackend.community.repository.PostLikeRepository;
+import pulleydoreurae.careerquestbackend.community.repository.PostRepository;
+import pulleydoreurae.careerquestbackend.community.repository.PostViewCheckRepository;
 import pulleydoreurae.careerquestbackend.common.community.domain.entity.Post;
 import pulleydoreurae.careerquestbackend.common.community.domain.entity.PostLike;
